util: split TypeName into separate file
In order to use this trick for other cases (registering classes that
correspond to standard library types), make this a separate util file.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4298, branch: user/amscanne/btf_types3/2
